moved swimming related route selection code from BasicRouteBehavior to SalmonBasicRouteBehavior.  Tested and worked as before the move.
git-svn-id: http://dminfo:8686/svn/repository/models/branches/ptm_development@2934 cdc4813b-4270-ec4f-936e-925f93a782c2

+++ src/ptm/DWR/DMS/PTM/Channel.java
@@ -461,7 +461,7 @@ public class Channel extends Waterbody{
   public void setChanDir(int chanDir){
 	  _chanDir = chanDir;
   }
-  public float getChanDir(){
+  public int getChanDir(){
 	  return _chanDir;
   }
